+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Bug Fixes
+
+* prevent transparent RenderTexture edge halos in the CRT display pass
+
 ## [1.1.4](https://github.com/Natteens/retropsx-urp/compare/v1.1.3...v1.1.4) (2026-09-05)